添加登陆后才能更新、删除的权限

+++ routes/api.js
@@ -18,21 +18,25 @@ router.post('/notes/add', function(req, res, next) {
     return res.send({status: 1, errorMsg: '请登录'});
   }
   
+  var uid = req.session.user.id;
   var note = req.body.note;
-  Note.create({text: note}).then( () => {
+
+  Note.create({text: note, uid: uid}).then( () => {
     res.send({status: 0})
   }).catch( () => {
     res.send({status: 1, errorMsg: '数据库出错'})    
   })
-  console.log('addd', note);
+  console.log('add', note);
 })
 
 router.post('/notes/edit', function(req, res, next) {
   if(!req.session.user){
     return res.send({status: 1, errorMsg: '请登录'});
   }
+
+  var uid = req.session.user.id;
   
-  Note.update({text: req.body.note}, {where: {id: req.body.id}}).then( () => {
+  Note.update({text: req.body.note}, {where: {id: req.body.id, uid: uid}}).then( () => {
     res.send({status: 0});
   }).catch( () => {
     res.send({status: 1, errorMsg: '数据库出错'})    
@@ -44,7 +48,9 @@ router.post('/notes/delete', function(req, res, next) {
     return res.send({status: 1, errorMsg: '请登录'});
   }
 
-  Note.destroy({where: {id: req.body.id}}).then( () => {
+  var uid = req.session.user.id;
+  
+  Note.destroy({where: {id: req.body.id, uid: uid}}).then( () => {
     res.send({status: 0})
   }).catch( () => {
     res.send({status: 1, errorMsg: '数据库出错'})
